+++ enclone_core/src/opt_d.rs
@@ -93,10 +93,10 @@ pub fn evaluate_d(
     // Align the V..J sequence on the contig to the reference concatenation.
 
     let mut seq_end = tig.len() - (jref.len() - jend);
-    if seq_end <= seq_start as usize {
+    if seq_end <= seq_start {
         seq_end = tig.len(); // bug fix for problem found by customer, couldn't reproduce internally
     }
-    let seq = &tig[seq_start as usize..seq_end];
+    let seq = &tig[seq_start..seq_end];
     let jref = &jref[0..jend];
     concat.extend(jref);
     let (ops, count) = align_to_vdj_ref(
